refactor: extract census's Network type into separate file
